+++ tests/Common/CommonColumnSchemaBuilderTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Tests\Common;
 
 use PHPUnit\Framework\TestCase;
+use Yiisoft\Db\Tests\Support\Stub\Column;
 use Yiisoft\Db\Tests\Support\TestTrait;
 
 use function array_shift;
@@ -24,10 +25,7 @@ abstract class CommonColumnSchemaBuilderTest extends TestCase
 
     protected function checkBuildString(string $expected, string $type, int|null $length, array $calls): void
     {
-        $db = $this->getConnection();
-
-        $schema = $db->getSchema();
-        $builder = $schema->createColumnSchemaBuilder($type, $length);
+        $builder = new Column($type, $length);
 
         foreach ($calls as $call) {
             $method = array_shift($call);
@@ -35,7 +33,5 @@ abstract class CommonColumnSchemaBuilderTest extends TestCase
         }
 
         $this->assertSame($expected, $builder->asString());
-
-        $db->close();
     }
 }
