+++ tests/Support/Stub/Column.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Tests\Support\Stub;
+
+use Yiisoft\Db\Schema\Builder\AbstractColumn;
+
+final class Column extends AbstractColumn
+{
+}
