--- tests/Support/Stub/ColumnSchemaBuilder.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Db\Tests\Support\Stub;
-
-use Yiisoft\Db\Schema\AbstractColumnSchemaBuilder;
-
-final class ColumnSchemaBuilder extends AbstractColumnSchemaBuilder
-{
-}
